app.py
from flask import Flask
from routes.Login.Login import auth_bt 
from routes.CreateAccount.CreateAccount import Cre_acc
from models.Sql_Tables import db 
from routes.CreateAccount.Functions.Transdatato_Db import create_acc
from routes.BMI.Bmi_Cal import Bmi_auth
from routes.Dashboard.Dashboard import dashboard_bp
from routes.Welcome.hello import Hi_bp
from routes.Dashboard.Functions.Dataentry import Dashboard_Details_Entry
from routes.Workout.Email_Reminder import EmailReminder
from routes.Workout.Email_Reminder import Reminder_app
from routes.Predection.index import prd_bp
from routes.FitnessBot.Main import Cht_bp


from dotenv import load_dotenv
import os 
load_dotenv()

app = Flask(__name__)
Mysql_DB=os.getenv("Mysql_DB")
# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///users.db'

#this line for local
# app.config['SQLALCHEMY_DATABASE_URI'] = f'mysql+pymysql://root:{Mysql_DB}@localhost/fitness_tracker'


if os.environ.get("RENDER") or os.environ.get("PYTHONANYWHERE"):
    app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite:///site.db"
else:
    app.config['SQLALCHEMY_DATABASE_URI'] = f"mysql+pymysql://root:{Mysql_DB}@localhost/fitness_tracker"


app.secret_key=os.getenv("SECRET_KEY")
db.init_app(app)
app.register_blueprint(Cht_bp)
app.register_blueprint(create_acc)
app.register_blueprint(dashboard_bp)
app.register_blueprint(auth_bt)
app.register_blueprint(Cre_acc)
app.register_blueprint(Bmi_auth)
app.register_blueprint(Hi_bp)
app.register_blueprint(Reminder_app)
app.register_blueprint(prd_bp)
# app.register_blueprint(Dashboard_Details_Entry)
# @app.before_request
# def check_access():
#     # This will run before every single request to your server
#     if Access(ip=request.remote_addr) == 0:
#         return render_template("Timeout.html")
with app.app_context():
    db.create_all() 
if __name__ == '__main__':

        # if os.environ.get('WERKZEUG_RUN_MAIN') == 'true' or not app.debug:
        # EmailReminder(timelimit=0)
    app.run(debug=True)
    