requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
python-dotenv
pandas
xgboost
ollama
requests